+++ sketch_oxide/src/similarity/mod.rs
@@ -22,6 +22,7 @@ pub mod minhash_lsh;
 pub mod oph;
 pub mod simhash;
 pub mod simhash_lsh;
+pub mod weighted_minhash;
 
 pub use bbit_minhash::BBitMinHash;
 pub use minhash::MinHash;
@@ -29,3 +30,4 @@ pub use minhash_lsh::MinHashLsh;
 pub use oph::OnePermutationHash;
 pub use simhash::SimHash;
 pub use simhash_lsh::{hamming_distance, SimHashLsh};
+pub use weighted_minhash::WeightedMinHash;